layers: Move WSI state to dedicated file
