+++ BUILD.gn
@@ -298,6 +298,8 @@ vvl_sources = [
   "layers/state_tracker/vertex_index_buffer_state.h",
   "layers/state_tracker/video_session_state.cpp",
   "layers/state_tracker/video_session_state.h",
+  "layers/state_tracker/wsi_state.cpp",
+  "layers/state_tracker/wsi_state.h",
   "layers/stateless/sl_buffer.cpp",
   "layers/stateless/sl_cmd_buffer.cpp",
   "layers/stateless/sl_cmd_buffer_dynamic.cpp",
